After decompressing the tarball, setup the working directory
This isn't a bare git repo, it's a proper working directory and some
plugins might assume the code is checked out. We'll assume there is an
`origin/main` branch and reset to that. The checkout phase comes shortly
after this hook, and it will pull extra objects then reset to the commit
being built.

Also reset submodules, just to ensure they all exist on disk too. In
case something assumes they're there.